CR: coord; add TODOs and comments
